Correção de erro ao tentar excluir um User que está registrado na tabela History

+++ src/main/java/io/github/arthoura/rest/UserResource.java
@@ -1,6 +1,8 @@
 package io.github.arthoura.rest;
 
+import io.github.arthoura.domain.model.Parking;
 import io.github.arthoura.domain.model.User;
+import io.github.arthoura.domain.repository.ParkingRepository;
 import io.github.arthoura.domain.repository.UsersRepository;
 import io.github.arthoura.rest.Dto.CreateUserRequest;
 import io.quarkus.hibernate.orm.panache.PanacheQuery;
@@ -10,20 +12,26 @@ import jakarta.annotation.security.PermitAll;
 import jakarta.annotation.security.RolesAllowed;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
+import jakarta.validation.ConstraintViolationException;
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+import org.jboss.resteasy.spi.UnhandledException;
+import org.postgresql.util.PSQLException;
+
 import java.util.Optional;
 
 @Path("/user")
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
 public class UserResource {
+    private ParkingRepository parkingRepository;
     private UsersRepository repository;
 
     @Inject
-    public UserResource(UsersRepository repository) {
+    public UserResource(UsersRepository repository, ParkingRepository parkingRepository) {
         this.repository = repository;
+        this.parkingRepository = parkingRepository;
     }
 
     @POST
@@ -34,6 +42,12 @@ public class UserResource {
         if(userConsultByCpf.isPresent()){
             return Response.status(Response.Status.CONFLICT)
                     .entity("{\"message\": \"Usuário com esse CPF já existe.\"}")
+                    .build();
+        }
+        Optional<User> userConsultByEmail = repository.find("email", createUserRequest.getEmail()).firstResultOptional();
+        if(userConsultByEmail.isPresent()){
+            return Response.status(Response.Status.CONFLICT)
+                    .entity("{\"message\": \"Usuário com esse Email já existe.\"}")
                     .build();
         }
         User user = new User();
@@ -78,8 +92,14 @@ public class UserResource {
     public Response deleteUser(@PathParam("userId") Long id){
         User userById = repository.findById(id);
         if(userById != null){
-            repository.delete(userById);
-            return Response.status(Response.Status.NO_CONTENT).build();
+            Parking parking = parkingRepository.find("user", userById).firstResult();
+            if(parking == null) {
+                repository.delete(userById);
+                return Response.status(Response.Status.NO_CONTENT).build();
+            }
+                return Response.status(Response.Status.CONFLICT)
+                        .entity("{\"mensagem\":\"usuário não pode ser excluído no momento pois está dentro do estacionamento\"}")
+                        .build();
         }
         return Response.status(Response.Status.NOT_FOUND).build();
     }
